#define -> const unsinged int
--HG--
branch : 1.04

+++ src/Diff.hpp
@@ -308,7 +308,7 @@ namespace dtl {
       sesElemVec ses_v = ses.getSequence();
       int l_cnt  = 1;
       int length = distance(ses_v.begin(), ses_v.end());
-      int middle = 0;
+      uint middle = 0;
       bool isMiddle, isAfter;
       isMiddle = isAfter = false;
       elem e;
@@ -384,8 +384,8 @@ namespace dtl {
         // compose unified format hunk
         if (isAfter && !change.empty()) {
           sesElemVec_iter cit = it;
-          int cnt = 0;
-          for (int i=0;i<DTL_SEPARATE_SIZE;++i, ++cit) {
+          uint cnt = 0;
+          for (int i=0;i<(int)DTL_SEPARATE_SIZE;++i, ++cit) {
             if (cit->second.type == SES_COMMON) {
               ++cnt;
             }
@@ -400,7 +400,7 @@ namespace dtl {
             rotate(common[0].begin(), 
                    common[0].begin() + c0size - DTL_SEPARATE_SIZE, 
                    common[0].end());
-            for (int i=0;i<c0size-DTL_SEPARATE_SIZE;++i) {
+            for (int i=0;i<c0size-(int)DTL_SEPARATE_SIZE;++i) {
               common[0].pop_back();
             }
             a += c0size - DTL_SEPARATE_SIZE;
